fix(engine): harden fill engine against memory/resource issues at scale (#526) (#529)
* fix(engine): harden against overflow, poisoned connections, bad config, FK cycles (#526)

First batch of the #526 hardening audit — small, isolated correctness fixes:

- H1: widen the FlatFileGenerator row loop counter from int to long so row
  counts past Integer.MAX_VALUE cannot overflow into an infinite write.
- H2: when re-enabling constraints on a bulk-load connection throws, abort the
  connection so the pool discards it instead of returning a constraint-disabled,
  FK-skipping session to the next borrower. Shared restoreConstraints helper now
  covers both the worker body and the fillUnordered probe (which had the same bug).
- M1: reject batchSize < 1 in the DatabaseConfiguration compact constructor.
- M5: add a visited-set cycle guard to getAssociatedPrimaryKeyColumn so a
  circular FK schema warns and returns null instead of StackOverflowError.
- M3: enforce and document ZipfianIntegerGenerator.MAX_SIZE so an oversized size
  fails fast instead of OOM-ing on a multi-gigabyte cumulative table.

Tests: exact-row-count guard, batchSize rejection, FK-cycle termination,
Zipfian size cap, and a Testcontainers test asserting a failed constraint
restore evicts the connection rather than poisoning the pool.


* fix(engine): bound parallel transactions/scheduling and guard restore/rollback (#526)

Second batch of the #526 hardening audit — the parallel/bulk scheduling and
transaction items plus the low-priority cleanups:

- M2: the parallel/bulk path now defaults CONNECTION_DEFAULT to a bounded
  everyNBatches(64) commit instead of perTable, so a very large partition no
  longer holds one unbounded server-side transaction (WAL/undo growth, lock
  accumulation). An explicitly configured strategy (incl. perTable) is honored.
- M4: parallel and unordered-bulk submission now run through runWithBackpressure,
  which keeps at most ~2x poolSize tasks in flight via an ExecutorCompletionService
  instead of invokeAll-ing every task at once, so queued FutureTasks/Futures scale
  with the pool size rather than total partition/task count.
- L1: visualizeGraph skips building the DOT string and URL-encoded link unless the
  consuming log level is enabled, and skips the (encode-tripled) link entirely for
  schemas whose DOT exceeds MAX_GRAPH_LINK_CHARS.
- L2: autocommit restore moved out of finally; a rollback or restore failure is now
  attached via addSuppressed instead of masking the original cause, and a connection
  whose autocommit can't be restored is aborted so a pool discards it.
- L3: explicit clearBatch() after every executeBatch(), making the O(batchSize)
  in-flight bound explicit and guarding non-spec-compliant rewrite paths.
- L4: pre-size the BitStringGenerator StringBuilder. The array generators' boxing is
  intrinsic to their Byte[]/Integer[]/String[] value types and JDBC createArrayOf,
  so it is left as-is.

Tests: an H2-backed test asserting a failed autocommit restore aborts the
connection and surfaces (not masks) the failure, and a Postgres test asserting
every partition task runs when partitions far exceed worker threads. Full
bloviate-core suite (231) plus junit/testcontainers modules green.


---------

+++ bloviate-core/src/main/java/io/bloviate/file/FlatFileGenerator.java
@@ -104,7 +104,7 @@ public class FlatFileGenerator implements FileGenerator {
 
             printHeader(csvPrinter);
 
-            for (int i = 0; i < rows; i++) {
+            for (long i = 0; i < rows; i++) {
                 for (ColumnDefinition columnDefinition : columnDefinitions) {
                     csvPrinter.print(columnDefinition.dataGenerator().generateAsString());
                 }
